Install rxjs
