+++ package.json
@@ -17,6 +17,7 @@
     "react-redux": "^7.2.4",
     "redux": "^4.1.1",
     "redux-thunk": "^2.3.0",
+    "rxjs": "^7.3.0",
     "sass": "^1.38.2"
   },
   "devDependencies": {
