+++ src/services/Buttons/index.tsx
@@ -1,5 +1,4 @@
 import http from "../http-common";
-import type { NextApiRequest, NextApiResponse } from 'next'
 import ButtonData from "../types/button.type"
